+++ ChebyshevFilterGenerator/ChebyshevFilterGenerator/cFilterDesign.h
@@ -19,8 +19,8 @@ public:
 
 	// Public Methods
 	std::vector<std::complex<double>> getTransferFunction(bool bAnalog);
-	std::vector<float> getYAxis(bool bAnalog, bool bMag);
-	std::vector<float> getXAxis(bool bAnalog);
+	std::vector<double> getYAxis(bool bAnalog, bool bMag);
+	std::vector<double> getXAxis(bool bAnalog);
 
 	// Public Attributes
 
@@ -36,26 +36,26 @@ private:
 	double t_n(double& fFreq);
 
 	// Private Attributes
-	double m_fDiscretePass_rads;									// Discrete Edge Frequency (rad/s)
-	double m_fDiscreteStop_rads;									// Discrete Corner Frequency (rad/s)
-	double m_fOmegaPass_rads;										// Analog Edge Frequency (rad/s)
-	double m_fOmegaStop_rads;										// Analog Corner Frequency (rad/s)
-	double m_fRipplePass_ratio;										// Analog Ripple Passband (ratio)
-	double m_fRippleStop_ratio;										// Analog Ripple Stopband (ratio)
-	double fEpsilon;												// Epsilon for Analog Filter Design
-	double m_fNumerator_s;											// Numerator of Analog Transfer Function
+	double m_fDiscretePass_rads;										// Discrete Edge Frequency (rad/s)
+	double m_fDiscreteStop_rads;										// Discrete Corner Frequency (rad/s)
+	double m_fOmegaPass_rads;											// Analog Edge Frequency (rad/s)
+	double m_fOmegaStop_rads;											// Analog Corner Frequency (rad/s)
+	double m_fRipplePass_ratio;											// Analog Ripple Passband (ratio)
+	double m_fRippleStop_ratio;											// Analog Ripple Stopband (ratio)
+	double fEpsilon;													// Epsilon for Analog Filter Design
+	double m_fNumerator_s;												// Numerator of Analog Transfer Function
 	
-	int m_iSampleRate_Hz;											// Sample Rate (Hz)
-	int m_iOrder_N;													// Order of designed filter
+	int m_iSampleRate_Hz;												// Sample Rate (Hz)
+	int m_iOrder_N;														// Order of designed filter
 
-	std::vector<std::complex<double>> m_vfTransferFunction_s;		// Transfer Function vector of Analog Filter
-	std::vector<double> m_vfDenominator_s;							// Denominator of Analog Transfer Function
-	std::vector<float> m_vfMagnitude_s;								// Analog Filter Magnitude Response	
-	std::vector<float> m_vfMagnitude_z;								// Digital Filter Magnitude Response	
-	std::vector<float> m_vfPhase_s;									// Analog Filter Phase Response
-	std::vector<float> m_vfPhase_z;									// Digital Filter Phase Response
-	std::vector<float> m_vfX_s;										// X-axis of the Analog Magnitude	
-	std::vector<float> m_vfX_z;										// X-axis of the Digital Magnitude
+	std::vector<std::complex<double>> m_vfTransferFunction_s;			// Transfer Function vector of Analog Filter
+	std::vector<double> m_vfDenominator_s;								// Denominator of Analog Transfer Function
+	std::vector<double> m_vfMagnitude_s;								// Analog Filter Magnitude Response	
+	std::vector<double> m_vfMagnitude_z;								// Digital Filter Magnitude Response	
+	std::vector<double> m_vfPhase_s;									// Analog Filter Phase Response
+	std::vector<double> m_vfPhase_z;									// Digital Filter Phase Response
+	std::vector<double> m_vfX_s;										// X-axis of the Analog Magnitude	
+	std::vector<double> m_vfX_z;										// X-axis of the Digital Magnitude
 	
 };
